examples/cv_cancer_detection/cancer_model/requirements.txt
torch==1.8.1
torchvision==0.9.1
Pillow==8.2.0
numpy==1.19.5
scikit-learn==0.24.2
